+++ app/Http/Controllers/HomeController.php
@@ -20,16 +20,6 @@ class HomeController extends Controller
     }
 
     /**
-     * Get the changelog view.
-     *
-     * @return \Illuminate\View\View
-     */
-    public function changelog()
-    {
-        return view('changelog');
-    }
-
-    /**
      * Get the terms of service and privacy policy view.
      *
      * @return \Illuminate\View\View