+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## 1.x
 
+- dev
+  - Fix facebook ranks sync exception when there is no post
+  - Break line when post content is too long
+
 - 1.0.1 (2016-06-08)
   - Fix incorrect form error message
   - Fix incorrect user role check
-  - Support ipv6 if application is behind cloudflare
+  - Support ipv6 if application is behind CloudFlare
   - Support restrict user country
 
 - 1.0.0 (2016-06-01)