+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - dev
   - Fix facebook ranks sync exception when there is no post
   - Break line when post content is too long
+  - Optimize block words filter
 
 - 1.0.1 (2016-06-08)
   - Fix incorrect form error message